+++ dataquality-common/src/main/java/org/talend/dataquality/common/character/Acronym.java
@@ -1,7 +1,6 @@
 package org.talend.dataquality.common.character;
 
 import java.util.List;
-import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -30,7 +29,6 @@ public class Acronym {
     }
 
     public String transform(String str) {
-        StringBuilder sb = new StringBuilder();
         TokenizedString tokenizedString = new TokenizedString(str, delimiterPattern);
         List<String> tokens = tokenizedString.getTokens();
 
@@ -45,31 +43,32 @@ public class Acronym {
             start++;
         } while (firstApplied.isEmpty() && start < tokens.size());
 
-        sb.append(firstApplied);
-        if (separator != AcronymSeparator.KEEP_SPECIAL_CHARS) {
-            for (int i = start; i < tokens.size(); i++) {
-                String chars = abbrevMode.apply(tokens.get(i));
-                if (!chars.isEmpty()) {
-                    sb.append(separator.value).append(abbrevMode.apply(tokens.get(i)));
-                }
-            }
-        } else {
+        StringBuilder sb = new StringBuilder(firstApplied);
+
+        if (separator == AcronymSeparator.KEEP_SPECIAL_CHARS) {
             Pattern specialCharPattern = Pattern.compile(separator.getValue());
             List<String> separators = tokenizedString.getSeparators();
             int nextSeparator = tokenizedString.isStartingWithSeparator() ? 1 : 0;
+
             for (int i = 1; i < tokens.size(); i++) {
                 String chars = abbrevMode.apply(tokens.get(i));
                 if (!chars.isEmpty()) {
-
                     sb.append(getSpecialChars(separators.get(nextSeparator), specialCharPattern)).append(
                             abbrevMode.apply(tokens.get(i)));
                 }
                 nextSeparator++;
             }
-        }
+        } else {
+            for (int i = start; i < tokens.size(); i++) {
+                String chars = abbrevMode.apply(tokens.get(i));
+                if (!chars.isEmpty()) {
+                    sb.append(separator.getValue()).append(abbrevMode.apply(tokens.get(i)));
+                }
+            }
 
-        if (sb.length() > 0 && separator == AcronymSeparator.PERIOD) {
-            sb.append(separator.getValue());
+            if (sb.length() > 0 && AcronymSeparator.PERIOD == separator) {
+                sb.append(separator.getValue());
+            }
         }
 
         return sb.toString();
@@ -131,40 +130,6 @@ public class Acronym {
 
         public String getValue() {
             return value;
-        }
-    }
-
-    public enum AbbreviationMode {
-
-        FIRST_LETTERS_IGNORE_NUMERIC(false, false, StringHandler::firstCharIgnoreNumeric),
-        FIRST_UPPER_CASE_LETTERS_IGNORE_NUMERIC(false, true, StringHandler::firstUpperIgnoreNumeric),
-        ALL_UPPER_CASE_LETTERS_IGNORE_NUMERIC(false, true, StringHandler::allUpperIgnoreNumeric),
-        FIRST_LETTERS_KEEP_NUMERIC(true, false, StringHandler::firstCharKeepNumeric),
-        FIRST_UPPER_CASE_LETTERS_KEEP_NUMERIC(true, true, StringHandler::firstUpperKeepNumeric),
-        ALL_UPPER_CASE_LETTERS_KEEP_NUMERIC(true, true, StringHandler::allUpperKeepNumeric);
-
-        private final boolean keepDigits;
-
-        private final boolean isUpperCaseMode;
-
-        private final Function<String, String> function;
-
-        AbbreviationMode(boolean keepDigits, boolean isUpperCaseMode, Function<String, String> function) {
-            this.keepDigits = keepDigits;
-            this.isUpperCaseMode = isUpperCaseMode;
-            this.function = function;
-        }
-
-        public boolean keepsDigits() {
-            return keepDigits;
-        }
-
-        public boolean isUpperCaseMode() {
-            return isUpperCaseMode;
-        }
-
-        public String apply(String str) {
-            return function.apply(str);
         }
     }
 }
